+++ App/Viewer/ViewTable.cpp
@@ -122,6 +122,36 @@ int CViewTable::ShowTabBar(bool bShow)
     return 0;
 }
 
+int CViewTable::Screenslot(const QString &szFile)
+{
+    CFrmViewScroll* pScroll = qobject_cast<CFrmViewScroll*>(GetViewer(m_pTab->currentIndex()));
+    if(!pScroll) return -1;
+    CFrmViewer* pView = pScroll->GetViewer();
+    if(!pView) return -1;
+//    int x = 0;
+//    int y = 0;
+//    int w = pScroll->rect().width();
+//    int h = pScroll->rect().height();
+//    switch (pView->GetAdaptWindows()) {
+//    case CFrmViewer::ZoomToWindow:
+//    case CFrmViewer::KeepAspectRationToWindow:
+//        w = -1;
+//        h = -1;
+//        break;
+//    default:
+//        QPoint p = pScroll->rect().topLeft();
+//        p = pScroll->mapToGlobal(p);
+//        p = pView->mapFromGlobal(p);
+//        x = p.x();
+//        y = p.y();
+//        break;
+//    }
+    QImage img = pView->GrabImage();
+    if(img.save(szFile))
+        return 0;
+    return -2;
+}
+
 void CViewTable::SetAdaptWindows(CFrmViewer::ADAPT_WINDOWS aw, QWidget* p)
 {
     CFrmViewer* pView = nullptr;