+++ Viewer/FrmViewer.h
@@ -77,6 +77,8 @@ public:
     
     virtual int Load(QSettings &set);
     virtual int Save(QSettings &set);
+    
+    virtual QImage GrabImage(int x = 0, int y = 0, int w = -1, int h = -1);
 
 public Q_SLOTS:
     void slotSetDesktopSize(int width, int height);